requirements.txt
selenium==4.11.2
beautifulsoup4==4.12.2
Django==4.2.4
django-cors-headers==3.13.0
django-extensions==3.2.0
django-rest-multiple-models==2.1.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.20.0
numpy==1.25.2
yt-dlp==2023.7.6
ffmpeg-python==0.2.0
pytaglib==2.0.0
requests==2.31.0
psycopg2-binary==2.9.6